moved hostReview to features and widgets
