+++ src/features/HostReview/index.ts
@@ -1,0 +1,5 @@
+export { ReviewCardOffer } from "./ui/ReviewCardOffer/ReviewCardOffer";
+export { ReviewFullCard } from "./ui/ReviewFullCard/ReviewFullCard";
+export { ReviewMiniCard } from "./ui/ReviewMiniCard/ReviewMiniCard";
+
+export { UserCardFullInfo, ReviewOffer, UserCardInfo } from "./model/types/hostReview";
