+++ src/features/HostReview/model/types/hostReview.ts
@@ -10,3 +10,13 @@ export interface UserCardInfo {
 export interface UserCardFullInfo extends UserCardInfo {
     textReview: string;
 }
+
+export interface ReviewOffer {
+    id: number;
+    author: string;
+    authorAvatar: string;
+    rating: number;
+    title: string;
+    textReview: string;
+    date: Date;
+}
