+++ src/features/HostReview/ui/ReviewFullCard/ReviewFullCard.module.scss
@@ -2,7 +2,6 @@
     border: 1px solid var(--accent-elements);
     border-radius: 3px;
     max-width: 500px;
-    width: 100%;
     min-height: 100px;
     display: flex;
     flex-direction: column;
