+++ src/features/HostReview/ui/ReviewMiniCard/ReviewMiniCard.module.scss
