--- src/modules/ReviewAboutOffers/model/types/reviewAboutOffers.ts
@@ -1,9 +1,0 @@
-export interface ReviewOffer {
-    id: number;
-    author: string;
-    authorAvatar: string;
-    rating: number;
-    title: string;
-    textReview: string;
-    date: Date;
-}
