+++ src/widgets/HostReview/index.ts
@@ -1,3 +1,2 @@
-export { ReviewMiniCard } from "./ui/ReviewMiniCard/ReviewMiniCard";
-
+export { ReviewAboutOffers } from "./ui/ReviewAboutOffers/ReviewAboutOffers";
 export { ReviewAboutVolunteers } from "./ui/ReviewAboutVolunteers/ReviewAboutVolunteers";
