+++ src/widgets/HostReview/ui/ReviewAboutOffers/ReviewAboutOffers.module.scss
