+++ src/widgets/HostReview/ui/ReviewAboutVolunteers/ReviewAboutVolunteers.module.scss
@@ -13,7 +13,7 @@
     &::-webkit-scrollbar-track {
         background: var(--color-white);
     }
-
+    
     &::-webkit-scrollbar-thumb {
         border-radius: var(--radius);
         background-color: var(--bg-accent);
